+++ django_svelte/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class DjangoSvelteConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "django_svelte"