+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 * This changelog!
 * Demo project into this package with appropriate tests and what have you.
+* Support for retrieving staticfiles from the `django-compressor` manifest after offline compression.
 ### Changed
 * Handling of CSS to dedicated templatetag which can appropriately be invoked in the `head` of the template.
 * Documentation to speak more directly to handling CSS in the two most common use cases.